requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiofiles==24.1.0
aiohappyeyeballs
aiohttp
aiolimiter==1.2.1
aiosignal==1.3.2
alembic==1.15.2
annotated-types
anthropic==0.49.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
apache-libcloud==3.8.0
argcomplete==3.6.2
argon2-cffi
argon2-cffi-bindings
aria2==0.0.1b0
arrow
asgiref==3.8.1
astor==0.8.1
asttokens
async-lru
attrs
azure-core==1.33.0
azure-identity==1.21.0
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
babel
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4
bitsandbytes==0.45.5
bleach
blessed==1.20.0
blinker==1.9.0
blis==0.7.11
boto3==1.38.3
botocore==1.38.3
Brotli
build==1.2.2.post1
cached-property
cachetools==5.5.2
catalogue==2.0.10
certifi
cffi
charset-normalizer
chroma-hnswlib==0.7.6
chromadb==0.6.3
circuitbreaker==2.1.3
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorama
coloredlogs==15.0.1
comm
confection==0.1.5
contourpy==1.3.2
coolname==2.2.0
cramjam==2.10.0
cryptography==44.0.2
cut-cross-entropy==25.1.1
cycler==0.12.1
cymem==2.0.11
databricks-sdk==0.47.0
datasets==3.6.0
debugpy
decorator
deepspeed
defusedxml
Deprecated==1.2.18
diffusers==0.33.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docker-pycreds==0.4.0
docstring_parser==0.16
durationpy==0.9
einops==0.8.0
en-core-web-sm
eval_type_backport==0.2.2
evaluate==0.4.3
exceptiongroup
executing
fastapi==0.115.11
fastjsonschema
filelock
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
fqdn
frozenlist
fschat==0.2.36
fsspec==2023.6.0
fuzzywuzzy==0.18.0
gitdb==4.0.12
GitPython==3.1.43
gmpy2
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==2.10.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
gpustat==1.1.1
gql==3.5.2
graphene==3.4.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
h2
hf-xet==1.1.2
hf_transfer==0.1.9
hjson
hpack
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx
huggingface-hub==0.32.2
humanfriendly==10.0
hyperframe
idna
importlib_metadata==8.5.0
importlib_resources
ipykernel
ipython
ipython_pygments_lexers
ipywidgets
isodate==0.7.2
isoduration
itsdangerous==2.2.0
jedi
jieba==0.42.1
Jinja2==3.1.5
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
json5
jsonpointer==2.1
jsonschema
jsonschema-specifications
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kiwisolver==1.4.8
kubernetes==32.0.1
langcodes==3.5.0
language_data==1.3.0
latex2mathml==3.77.0
Levenshtein==0.27.1
lightning-utilities==0.14.3
llm-foundry==0.19.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe
matplotlib==3.10.1
matplotlib-inline
mdurl==0.1.2
mistune
mlflow==2.21.3
mlflow-skinny==2.21.3
mmh3==5.1.0
modelscope==1.28.0
monotonic==1.6
mosaicml==0.30.0
mosaicml-cli==0.7.2
mosaicml-streaming==0.12.0
mpmath
msal==1.32.3
msal-extensions==1.3.1
msgpack
msgspec==0.19.0
multidict
multiprocess==0.70.16
murmurhash==1.0.12
mypy==1.16.0
mypy_extensions==1.1.0
nbclient
nbconvert
nbformat
nest_asyncio
networkx
nh3==0.2.21
ninja==1.11.1.4
nltk==3.9.1
nodeenv==1.9.1
nodejs-wheel-binaries==22.16.0
notebook
notebook_shim
numpy
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
oci==2.150.2
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
openai==1.75.0
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
optree
optype==0.10.0
orjson==3.10.15
overrides
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.250527
pandocfilters
paramiko==3.5.1
parso
pathspec==0.12.1
peft==0.10.0
pexpect
pickleshare
pillow==11.1.0
pkgutil_resolve_name
platformdirs
ply
posthog==3.19.1
preshed==3.0.9
prometheus_client
prompt_toolkit
propcache==0.3.0
proto-plus==1.26.1
protobuf==3.20.3
psutil==6.1.1
ptyprocess
pure_eval
py-cpuinfo
pyarrow==20.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11
pybind11_global
pycparser
pydantic
pydantic_core
Pygments
PyJWT==2.10.1
PyNaCl==1.5.0
pynvml==11.5.3
pyOpenSSL==24.3.0
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5_sip
pyright==1.1.401
PySocks
python-dateutil
python-dotenv==1.0.1
python-json-logger
python-Levenshtein==0.27.1
python-snappy==0.7.3
pytorch-ranger==0.1.1
pytz==2025.1
PyYAML
pyzmq
qtconsole
QtPy
questionary==2.1.0
RapidFuzz==3.12.2
referencing
regex
requests
requests-oauthlib==2.0.0
rfc3339_validator
rfc3986-validator
rich==13.9.4
rouge==1.0.1
rouge_score==0.1.2
rpds-py
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.12.0
safetensors
scikit-learn==1.6.1
scipy==1.15.2
scipy-stubs==1.15.3.0
Send2Trash
sentence-transformers==3.4.1
sentencepiece==0.2.0
sentry-sdk==2.27.0
setproctitle==1.3.5
setuptools==72.1.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
sip
six
slack_sdk==3.35.0
smart-open==7.1.0
smmap==5.0.2
sniffio
sounddevice==0.5.2
soupsieve==2.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
sqlparse==0.5.3
srsly==2.5.1
stack_data
starlette==0.46.0
svgwrite==1.4.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
termcolor==3.0.1
terminado
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2
together==1.4.1
tokenizers==0.20.3
toml
tomli
torch==2.6.0+cu124
torch-optimizer==0.3.0
torchaudio==2.6.0+cu124
torchmetrics==1.6.0
torchvision==0.21.0+cu124
tornado
tqdm
traitlets
transformers==4.46.1
triton==3.2.0
trl==0.18.0
typeguard==4.4.2
typer==0.15.2
types-python-dateutil
types-pytz==2025.2.0.20250516
typing-inspection
typing_extensions
typing_utils
tyro==0.9.22
tzdata==2025.1
unsloth
unsloth_zoo==2025.5.10
uri-template
uritemplate==4.1.1
urllib3
uvicorn==0.34.0
uvloop==0.21.0
validators==0.34.0
wandb==0.18.7
wasabi==1.1.3
watchfiles==1.0.4
wavedrom==2.0.3.post3
wcwidth
weasel==0.4.1
webcolors
webencodings
websocket-client
websockets==11.0.3
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension
wrapt==1.17.2
xformers==0.0.30
xxhash==3.5.0
yarl
zipp
zstandard
zstd==1.5.6.7
